Fixes #24420 - CV Publish for modules
This commit provides the ability to publish content views on
repositories holding module streams and defaults. The main complication
is the implicit filtering on module streams. i.e. We do not copy module
streams that do not have any packages associated to them.

More details:
The publish algorithm works as follows
1) Packages are first filtered
2) Module Streams/Defaults are copied over from source repo to
the destination repo
3) Module Streams/Defaults without any packages/artifacts in the
destination repo are purged.