Fixes #28037 - Incremental Update additional repos
This commit handles incremental update with additional_repos parameters.
Content from one repo can depend on content from another repo, this is
especially true with RHEL 8. This causes a problem during depsolving.
For example if we have a module/errata in one repo pointing packages in
another repo, we'd like to ideally copy the dependencies across both
repositories in the same operation.
In the incremental update case for example the user may request a
certain errata to be copied over along with their dependencies.
To address this Pulp 2.21 updated their copy api to accept
"additional repos" (https://pulp.plan.io/issues/5067.)

This commit sets up the appropriate mapping for this operation. So if
the content view has depsolve turned on or if its incremental katello
will create a 'repository_mapping' which maps source repo to a
destination repo. The entries in this hash consist of all the
repositories in the content view version minus the main one.
For example
CVV1 - repos A,B
On incremental update CVV1 -> CVV 1.1
The copy operation for content in repo A will  have {b:dest-b} as the
"additional_repos" aka "repository mapping".
This mapping is then passed to pulp copy api.

This change has a couple of complexities. We need all the destination
repositories to be created before we call the copy operations with the
addition_repos mapping. To handle this a new CreateRepos plan was
created and the CloneToVersion plan was modified to accept a
destination repo. The CloneToVersion call now expects the destination
repo to have been created before copying.
Changes were made to both IncrementalUpdate and Publish plans to
accomodate the new structure.

In addition we need ensure that the errata is not purged from the
destination repo if the errata artifacts are available in one of the
dependent repositories. i.e. If a repo has only errata and the rpms
belonging to this errata are provided by a different dependent repo we
need to be able to handle that case and look at the dependent repos and
ascertain that the errata is satified.