Fixes #33451 - use upsert all for some content indexing
to avoid duplicate key errors.  Upsert all is only used for the inital
unit creation (with just the pulp_id), and creation of errata sub-items.
We could take this further and increase performance even more, but
when doing so this change became very large and complicated and
out of the scope of the original problem