BZ #2032098 - Incremental publish content view doesn't copy any contents
Fixes #34145 - Errata href not updated (#9854)

Like any other content in Pulp 3, advisories are immutable. So when
there is some change to it, a new one is created, it is then
associated to a repository as a part of a new repository version and
the old one is unassociated and becomes orphaned.

There are some advisory's fields that Katello doesn't store in the
database but Pulp does, such as version, pushcount, fromstr, rights
etc. If any of those fields is changed after syncing repo, Katello
can't detect the change and it is still referring to the orphaned
hrefs.

When user publishing an incremental content view, Katello will use the
orphaned hrefs to copy advisories from source repo to the destination
repo. If the orphaned advisories are removed by Pulp then nothing will
be copied.

(cherry picked from commit 7961da661073d22c9ae8e3f320186777aa85aa8b)