BZ #2058654 - Fixes #34660: Reload old recurring logic when updating sync plan
A change in Rails 6.0.4+ results in the old recurring logic retaining
a reference to the sync plan such that when the recurring logic is
cancelled, which triggers a save, that save resets the relationship
of sync plan to recurring logic. The resulting sync plan winds up
related to the old recurring logic instead of the new.

(cherry picked from commit 828e4f05e7facca2bdf9506c7885a86ae462723d)